+++ frontend/src/main/resources/webapp/src/app/main/authors/authors.module.ts
@@ -8,6 +8,12 @@ import {OntimizeWebModule} from 'ontimize-web-ngx';
 import {SharedModule} from "../../shared/shared.module";
 import {OpinionsModule} from "../opinions/opinions.module";
 
+import { AuthorTableComponent } from './authors-home/author-table/author-table.component';
+import { AuthorCardMediumComponent } from './cards/author-card-medium/author-card-medium.component';
+import { AuthorCardSmallComponent } from './cards/author-card-small/author-card-small.component';
+import { AuthorGridComponent } from './authors-home/author-grid/author-grid.component';
+
+
 @NgModule({
   imports: [
     CommonModule,
@@ -18,7 +24,12 @@ import {OpinionsModule} from "../opinions/opinions.module";
   ],
   declarations: [
     AuthorsHomeComponent,
-    AuthorsDetailComponent
+    AuthorsDetailComponent,
+    AuthorGridComponent,
+    AuthorTableComponent,
+    AuthorCardMediumComponent,
+    AuthorCardSmallComponent,
+
   ]
 })
 export class AuthorsModule {