+++ frontend/src/main/resources/webapp/src/app/main/publishers/publisher.ts
@@ -1,0 +1,6 @@
+export interface Publisher {
+    publisher_id: number,
+    publisher_name: string,
+    publisher_logo?:string,
+    publisher_desc:string,
+}