+++ frontend/src/main/resources/webapp/src/app/main/publishers/publishers.module.ts
@@ -1,12 +1,16 @@
-import {NgModule} from '@angular/core';
-import {CommonModule} from '@angular/common';
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
 
-import {OntimizeWebModule} from "ontimize-web-ngx";
-import {PublishersRoutingModule} from './publishers-routing.module';
+import { OntimizeWebModule } from "ontimize-web-ngx";
+import { PublishersRoutingModule } from './publishers-routing.module';
 import {PublishersHomeComponent} from "./publishers-home/publishers-home.component";
 import {PublishersDetailComponent} from "./publishers-detail/publishers-detail.component";
 import {SharedModule} from "../../shared/shared.module";
 import {OpinionsModule} from "../opinions/opinions.module";
+import {PublisherCardSmallComponent} from "./cards/publisher-card-small/publisher-card-small.component";
+import {PublisherCardMediumComponent} from "./cards/publisher-card-medium/publisher-card-medium.component";
+import {BookCardSmallComponent} from "../books/cards/book-card-small/book-card-small.component";
+import {BookCardMediumComponent} from "../books/cards/book-card-medium/book-card-medium.component";
 
 @NgModule({
   imports: [
@@ -16,10 +20,16 @@ import {OpinionsModule} from "../opinions/opinions.module";
     SharedModule,
     OpinionsModule
   ],
+    exports: [
+        PublisherCardSmallComponent,
+        PublisherCardMediumComponent
+    ],
 
-  declarations: [
-      PublishersHomeComponent,
-      PublishersDetailComponent
-  ]
+    declarations: [
+        PublishersHomeComponent,
+        PublishersDetailComponent,
+        PublisherCardSmallComponent,
+        PublisherCardMediumComponent
+    ]
 })
 export class PublishersModule { }